refactor: add plugin facades for printer and comment deps
