+++ src/plugin/src/component-providers/default-plugin-component-dependencies.js
@@ -1,9 +1,9 @@
 import { identifierCaseOptions } from "gamemaker-language-semantic/identifier-case/options.js";
 
-import { handleComments, printComment } from "../comments/comment-printer.js";
+import { handleComments, printComment } from "../comments/public-api.js";
 import { LogicalOperatorsStyle } from "../options/logical-operators-style.js";
-import { gmlParserAdapter } from "../parsers/gml-parser-adapter.js";
-import { print } from "../printer/print.js";
+import { gmlParserAdapter } from "../parsers/index.js";
+import { print } from "../printer/index.js";
 
 export const defaultGmlPluginComponentDependencies = Object.freeze({
     gmlParserAdapter,
