+++ src/plugin/src/printer/index.js
@@ -1,0 +1,6 @@
+// Public facade for printer utilities used by the Prettier plugin.
+//
+// Keeping the exports centralized here allows external consumers (including
+// sibling workspaces) to rely on a stable module path instead of importing
+// files from the internal directory layout.
+export { print } from "./print.js";
