+++ src/plugin/test/default-plugin-component-dependencies.test.js
@@ -2,12 +2,9 @@ import assert from "node:assert/strict";
 import test from "node:test";
 
 import { defaultGmlPluginComponentDependencies } from "../src/component-providers/default-plugin-component-dependencies.js";
-import { gmlParserAdapter } from "../src/parsers/gml-parser-adapter.js";
-import { print } from "../src/printer/print.js";
-import {
-    handleComments,
-    printComment
-} from "../src/comments/comment-printer.js";
+import { gmlParserAdapter } from "../src/parsers/index.js";
+import { print } from "../src/printer/index.js";
+import { handleComments, printComment } from "../src/comments/public-api.js";
 import { identifierCaseOptions } from "gamemaker-language-semantic/identifier-case/options.js";
 import { LogicalOperatorsStyle } from "../src/options/logical-operators-style.js";
